@@ -9,7 +9,7 @@ import android.net.NetworkInfo;
 import android.net.Uri;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ListView;
@@ -21,7 +21,9 @@ import java.util.List;
 public class MainActivity extends AppCompatActivity
         implements LoaderManager.LoaderCallbacks<List<News>> {
 
-    private static final String guardianURL = "https://content.guardianapis.com/search?tag=environment/recycling&api-key=test&format=json&show-tags=contributor";
+    private static final String guardianURL =
+            "https://content.guardianapis.com/search?tag=environment/recycling&api-key=" +
+                    "5d8383be-7d04-4c83-ada9-fd27b9d3c273&format=json&show-tags=contributor";
 
     private int mInterval = 120000; // 2 minutes default update rate
     private Handler mHandler;
@@ -77,17 +79,17 @@ public class MainActivity extends AppCompatActivity
         // Create a new adapter that takes an empty list of news as input
         mAdapter = new NewsAdapter(this, new ArrayList<News>());
 
-        // Find the {@link ListView} object in the view hierarchy of the {@link Activity}.
-        // There should be a {@link ListView} with the view ID called list, which is declared in the
+        // Find the ListView object in the view hierarchy of the Activity
+        // There should be a ListView with the view ID called list, which is declared in the
         // news_list.xml layout file.
-        ListView listView = (ListView) findViewById(R.id.list);
+        ListView listView = findViewById(R.id.list);
 
 
-        mEmptyStateTextView = (TextView) findViewById(R.id.empty_view);
+        mEmptyStateTextView = findViewById(R.id.empty_view);
         listView.setEmptyView(mEmptyStateTextView);
 
-        // Make the {@link ListView} use the {@link NewsAdapter} we created above, so that the
-        // {@link ListView} will display list items for each {@link News} in the list.
+        // Make the ListView use the NewsAdapter we created above, so that the
+        // ListView will display list items for each News in the list.
         listView.setAdapter(mAdapter);
 
         // Get a reference to the ConnectivityManager to check state of network connectivity
@@ -101,10 +103,6 @@ public class MainActivity extends AppCompatActivity
         if (networkInfo != null && networkInfo.isConnected()) {
             // Get a reference to the LoaderManager, in order to interact with loaders.
             LoaderManager loaderManager = getLoaderManager();
-
-            // Initialize the loader. Pass in the int ID constant defined above and pass in null for
-            // the bundle. Pass in this activity for the LoaderCallbacks parameter (which is valid
-            // because this activity implements the LoaderCallbacks interface).
             loaderManager.initLoader(NEWS_LOADER_ID, null, this);
         } else {
             // Otherwise, display error
@@ -123,13 +121,15 @@ public class MainActivity extends AppCompatActivity
                 News currentNews = mAdapter.getItem(position);
 
                 // Convert the String URL into a URI object (to pass into the Intent constructor)
-                Uri newsUri = Uri.parse(currentNews.getUrl());
+                if (currentNews != null) {
+                    Uri newsUri = Uri.parse(currentNews.getUrl());
 
-                // Create a new intent to view the news URI
-                Intent websiteIntent = new Intent(Intent.ACTION_VIEW, newsUri);
+                    // Create a new intent to view the news URI
+                    Intent websiteIntent = new Intent(Intent.ACTION_VIEW, newsUri);
 
-                // Send the intent to launch a new activity
-                startActivity(websiteIntent);
+                    // Send the intent to launch a new activity
+                    startActivity(websiteIntent);
+                }
             }
         });
 
