@@ -1,11 +1,7 @@
 package com.arby.recyclingnews;
 
-/**
- * Created by rares on 23.08.2017.
- */
-
 import android.app.Activity;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -18,10 +14,6 @@ import java.util.ArrayList;
 public class NewsAdapter extends ArrayAdapter<News> {
 
     public NewsAdapter(Activity context, ArrayList<News> adapter) {
-        // Here, we initialize the ArrayAdapter's internal storage for the context and the list.
-        // the second argument is used when the ArrayAdapter is populating a single TextView.
-        // Because this is a custom adapter for two TextViews and an ImageView, the adapter is not
-        // going to use this second argument, so it can be any value. Here, we used 0.
         super(context, 0, adapter);
     }
 
@@ -35,7 +27,7 @@ public class NewsAdapter extends ArrayAdapter<News> {
                     R.layout.list_item, parent, false);
         }
 
-        // Get the {@link News} object located at this position in the list
+        // Get the News object located at this position in the list
         final News currentNews = getItem(position);
 
         String title = currentNews.getTitle();
@@ -43,16 +35,16 @@ public class NewsAdapter extends ArrayAdapter<News> {
         String date = currentNews.getDate();
         String section = currentNews.getSection();
 
-        TextView titleView = (TextView) listItemView.findViewById(R.id.title);
+        TextView titleView = listItemView.findViewById(R.id.title);
         titleView.setText(title);
 
-        TextView authorView = (TextView) listItemView.findViewById(R.id.author);
+        TextView authorView = listItemView.findViewById(R.id.author);
         authorView.setText(author);
 
-        TextView sectionView = (TextView) listItemView.findViewById(R.id.section);
+        TextView sectionView = listItemView.findViewById(R.id.section);
         sectionView.setText(section);
 
-        TextView dateView = (TextView) listItemView.findViewById(R.id.date);
+        TextView dateView = listItemView.findViewById(R.id.date);
         dateView.setText(date);
 
         // Return the whole list item layout
